Update `ActiveEnumeration` to use `string.Equals` rather than `Utils.IsFileNameMatch` when the enumeration filter string does not have wildcards.

Times with scenario provided by @kyle-rader 

  | Baseline Times (sec) | PR Changes (sec)
-- | -- | --
  | 48.23 | 9.83
  | 41.16 | 9.37 
  | 44.49 | 10.76 
  | 41.18 | 10.35 
Average | 43.77 | 10.08